Updated to .NET 6 and improved Actions Workflows

+++ RICADO.Omron/OmronPLC.cs
@@ -115,9 +115,9 @@ namespace RICADO.Omron
 
         public string ControllerVersion => _controllerVersion;
 
-        public ushort MaximumReadWordLength => _plcType == enPLCType.CP1 ? 499 : 999;
+        public ushort MaximumReadWordLength => _plcType == enPLCType.CP1 ? (ushort)499 : (ushort)999;
 
-        public ushort MaximumWriteWordLength => _plcType == enPLCType.CP1 ? 496 : 996;
+        public ushort MaximumWriteWordLength => _plcType == enPLCType.CP1 ? (ushort)496 : (ushort)996;
 
         #endregion
